Add files via upload
This is a simple R program to crawl, merge and interpolate the daily closing prices of stock indices and their constituents. The program requires the "reticulate" package in R and python software to be installed. It is intended for R users to access financial data online in R software, for other functions see https://akshare.xyz/.